Enable per-memcg accounting of virtual memory
Based on Andrea Righi's patches in 2008:

    https://lkml.org/lkml/2008/6/9/490

This enables each control group to respect its own overcommit behavior
and return ENOENT in response to requests for virtual memory.  Some
amount of refactoring should be extended to nommu.c, which contains a
near-duplicate of the same code.

Extremely unsatisfying is the treatment of "charged" in the mm_struct,
which exists because during process exit the mm_struct's owning task
(via ->owner) is NULL-ified by the time the amount of virtual memory
to be freed is computed, which is currently coupled to the operator
that also frees that virtual memory.  I don't think this problem
existed when Andrea Righi posted his patches back then, or this patch
didn't work right, because without taking into consideration a NULL
owner of a mm_struct in exit.c I do a NULL dereference more or less
immediately.

Also dissatisfying is the treatment of SysV shmem segments.  In this
case, it seems like the right aesthetic is more straightforward: SysV
shmem should be moved around between cgroups just like processes,
since SysV shared memory exists independent of any process.

I'm also pretty sure this breaks some non-memcg enabled builds, but
right now this patch is only at a stage for discussion anyway.

Signed-off-by: Dan Farina <drfarina@acm.org>